CP6-94 mwalib added to docker and metadata file reader finds all input files
mwalib library added to the docker file.
Added constructor to MetadataFileReader (incomplete).
Altered Common.hpp to allow compilation (changed unsigned to unsigned long long).
MetadataFileReader constructor takes in an AppConfig and finds the metadata file and voltage data files for the specified observation and start time.

+++ src/Common.hpp
@@ -9,8 +9,8 @@
 // Entered as command line arguments
 struct AppConfig {
 	std::string inputDirectoryPath;
-	unsigned observationID;
-	unsigned signalStartTime;
+	unsigned long long observationID;
+	unsigned long long signalStartTime;
 	std::string invPolyphaseFilterPath;
 	std::string outputDirectoryPath;
 };
